Review feedback.

+++ pg_search/src/scan/batch_scanner.rs
@@ -155,8 +155,8 @@ impl Scanner {
     }
 
     /// Returns the estimated number of rows that will be produced by this scanner.
-    pub fn estimated_rows(&self) -> usize {
-        self.search_results.estimated_doc_count() as usize
+    pub fn estimated_rows(&self) -> u64 {
+        self.search_results.estimated_doc_count()
     }
 
     fn try_get_batch_ids(&mut self) -> Option<(SegmentOrdinal, Vec<f32>, Vec<u32>)> {
